sample_app/requirements.txt
fastapi
uvicorn
google-cloud-logging
python-multipart